chore: Upgraded dependency vitest => ^v1.0.1 across workspace in @observerly/fits.
chore: Upgraded dependency vitest => ^v1.0.1 across workspace in @observerly/fits. Includes reorganising tests into their namespaced modules.
