+++ package.json
@@ -29,6 +29,11 @@
       "main"
     ]
   },
+  "os": [
+    "darwin",
+    "linux"
+  ],
+  "packageManager": "pnpm@8.10.2",
   "scripts": {
     "dev": "vite",
     "build": "tsc --noEmit && vite build",
@@ -46,7 +51,7 @@
     "prettier": "^2.8.8",
     "tslib": "^2.6.1",
     "typescript": "^5.1.3",
-    "vite": "^4.4.7",
-    "vitest": "^0.33.0"
+    "vite": "^5.0.8",
+    "vitest": "^1.0.1"
   }
 }
