+++ src/header/__tests__/getFITSHeaders.spec.ts
@@ -6,16 +6,11 @@
 
 /*****************************************************************************************************************/
 
-import type { FITSHeader } from '../types'
-
 import { describe, expect, it, suite } from 'vitest'
 
-import {
-  getFITSHeaders,
-  parseFITSHeaderBlock,
-  parseFITSHeaderRow,
-  readFITSHeaderFromBlocks
-} from '..'
+import { type FITSHeader } from '../../types'
+
+import { getFITSHeaders } from '../'
 
 /*****************************************************************************************************************/
 
@@ -63,106 +58,6 @@ suite('@observerly/fits Header', () => {
       const NAXIS2 = headers.find((header: FITSHeader) => header.key === 'NAXIS2')
       expect(NAXIS2).toBeDefined()
       expect(NAXIS2).toMatchObject({ key: 'NAXIS2', value: 3672, comment: '' })
-    })
-  })
-
-  describe('FITS File Header Block Parsing', () => {
-    it('parseFITSHeaderBlock should be defined', () => {
-      expect(parseFITSHeaderBlock).toBeDefined()
-    })
-  })
-
-  describe('FITS File Header Row Parsing', () => {
-    it('parseFITSHeaderRow should be defined', () => {
-      expect(parseFITSHeaderRow).toBeDefined()
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow(
-        'SIMPLE  =                    T / file does conform to FITS standard'
-      )
-
-      expect(key).toBe('SIMPLE')
-      expect(value).toBe(true)
-      expect(comment).toBe('file does conform to FITS standard')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow(
-        'BITPIX  =                    8 / number of bits per data pixel'
-      )
-
-      expect(key).toBe('BITPIX')
-      expect(value).toBe(8)
-      expect(comment).toBe('number of bits per data pixel')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow(
-        'NAXIS   =                    0 / number of array dimensions or data axes'
-      )
-
-      expect(key).toBe('NAXIS')
-      expect(value).toBe(0)
-      expect(comment).toBe('number of array dimensions or data axes')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow(
-        'NAXIS   =                    0 / number of array dimensions or data axes'
-      )
-
-      expect(key).toBe('NAXIS')
-      expect(value).toBe(0)
-      expect(comment).toBe('number of array dimensions or data axes')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow('EXTEND  =                    T')
-
-      expect(key).toBe('EXTEND')
-      expect(value).toBe(true)
-      expect(comment).toBe('')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow(
-        "EPOCH   =             'Julian' / epoch of observation"
-      )
-
-      expect(key).toBe('EPOCH')
-      expect(value).toBe('Julian')
-      expect(comment).toBe('epoch of observation')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow('CONTINUE Continued value')
-
-      expect(key).toBe('CONTINUE')
-      expect(value).toBe('Continued value')
-      expect(comment).toBe('Continued value')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow('COMMENT An unspecified comment')
-
-      expect(key).toBe('COMMENT')
-      expect(value).toBe('An unspecified comment')
-      expect(comment).toBe('An unspecified comment')
-    })
-
-    it('parseFITSHeaderRow should return the correct key, value and comment', () => {
-      const { key, value, comment } = parseFITSHeaderRow('HISTORY An unspecified record')
-
-      expect(key).toBe('HISTORY')
-      expect(value).toBe('An unspecified record')
-      expect(comment).toBe('An unspecified record')
-    })
-  })
-
-  describe('FITS File Header Blocks Reading', () => {
-    it('readFITSHeaderFromBlocks should be defined', () => {
-      expect(readFITSHeaderFromBlocks).toBeDefined()
     })
   })
 })
