+++ src/header/__tests__/readFITSHeaderRow.spec.ts
@@ -8,19 +8,14 @@
 
 import { describe, expect, it, suite } from 'vitest'
 
-import { getExcessByteSize } from '../utilities'
+import { readFITSHeaderFromBlocks } from '../'
 
 /*****************************************************************************************************************/
 
-suite('@observerly/fits Utilities', () => {
-  describe('FITS File Parsing Utilities', () => {
-    it('getExcessByteSize should be defined', () => {
-      expect(getExcessByteSize).toBeDefined()
-    })
-
-    it('getExcessByteSize should return the correct number of excess bytes', () => {
-      const excess = getExcessByteSize(2880, 68)
-      expect(excess).toBe(2880 - 68)
+suite('@observerly/fits Header', () => {
+  describe('FITS File Header Blocks Reading', () => {
+    it('readFITSHeaderFromBlocks should be defined', () => {
+      expect(readFITSHeaderFromBlocks).toBeDefined()
     })
   })
 })
